Update hydrant display on zoom as well as pan
The app was using the 'dragend' event on the map to listen for
changes that should prompt the display of markers, but that event
only triggers when the map is panned. Changing to the 'idle' event
ensures that the display is updated on both zoom and pan instead
of requiring users to slightly pan the map after zooming to get
new data.

This has the side effect of making the intial call to
addMarkersAround unnecessary, as the idle event occurs when the
map is initially loaded, so I've removed that call.